historic-artwork-forgery-detection/requirements.txt
numpy
pandas
Pillow
tqdm
scikit-learn
matplotlib
seaborn
torch
torchvision
tensorflow
tensorflow-addons
opencv-python
streamlit

